adds block module
