+++ src/tasks/block/mod.rs
@@ -1,0 +1,8 @@
+//! block holds the tasks responsible for block environment configuration, 
+//! preparation, and simulation.
+
+/// Block simulation task for building blocks.
+pub mod sim;
+
+/// Block environment configuration for the builder.
+pub mod cfg;
