+++ src/tasks/mod.rs
@@ -1,6 +1,3 @@
-/// Block creation task
-pub mod block;
-
 /// Bundle poller task
 pub mod bundler;
 
@@ -15,3 +12,6 @@ pub mod submit;
 
 /// Tx polling task
 pub mod tx_poller;
+
+/// Block simulation and environment
+pub mod block;
